+++ libs/cpp_dbc/include/cpp_dbc/drivers/driver_sqlite.hpp
@@ -84,7 +84,7 @@ namespace cpp_dbc
             bool isNull(const std::string &columnName) override;
 
             std::vector<std::string> getColumnNames() override;
-            int getColumnCount() override;
+            size_t getColumnCount() override;
             void close() override;
 
             // BLOB support methods
