+++ libs/cpp_dbc/test/test_mysql_real_full_join.cpp
@@ -253,7 +253,7 @@ TEST_CASE("MySQL FULL JOIN operations (emulated)", "[mysql_real_full_join]")
             {6, "Eva Wilson", std::nullopt, std::nullopt},
             {7, "Frank Miller", std::nullopt, std::nullopt}};
 
-        int rowCount = 0;
+        size_t rowCount = 0;
         while (rs->next())
         {
             REQUIRE(rowCount < expectedResults.size());
@@ -317,7 +317,7 @@ TEST_CASE("MySQL FULL JOIN operations (emulated)", "[mysql_real_full_join]")
             {106, "Keyboard", std::nullopt, std::nullopt},
             {107, "Mouse", std::nullopt, std::nullopt}};
 
-        int rowCount = 0;
+        size_t rowCount = 0;
         while (rs->next())
         {
             REQUIRE(rowCount < expectedResults.size());
@@ -385,7 +385,7 @@ TEST_CASE("MySQL FULL JOIN operations (emulated)", "[mysql_real_full_join]")
             {"John Doe", "Laptop", 1, 999.99},
             {"John Doe", "Tablet", 2, 599.98}};
 
-        int rowCount = 0;
+        size_t rowCount = 0;
         while (rs->next())
         {
             REQUIRE(rowCount < expectedResults.size());
@@ -459,7 +459,7 @@ TEST_CASE("MySQL FULL JOIN operations (emulated)", "[mysql_real_full_join]")
             {"Eva Wilson", std::nullopt, std::nullopt, std::nullopt},
             {"Frank Miller", std::nullopt, std::nullopt, std::nullopt}};
 
-        int rowCount = 0;
+        size_t rowCount = 0;
         while (rs->next())
         {
             REQUIRE(rowCount < expectedResults.size());
