+++ libs/cpp_dbc/test/test_mysql_real_left_join.cpp
@@ -1,5 +1,5 @@
 /**
- 
+
  * Copyright 2025 Tomas R Moreno P <tomasr.morenop@gmail.com>. All Rights Reserved.
  *
  * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
@@ -248,7 +248,7 @@ TEST_CASE("MySQL LEFT JOIN operations", "[mysql_real_left_join]")
             {6, "Eva Wilson", std::nullopt, std::nullopt},
             {7, "Frank Miller", std::nullopt, std::nullopt}};
 
-        int rowCount = 0;
+        size_t rowCount = 0;
         while (rs->next())
         {
             REQUIRE(rowCount < expectedResults.size());
@@ -299,7 +299,7 @@ TEST_CASE("MySQL LEFT JOIN operations", "[mysql_real_left_join]")
             {"John Doe", "Laptop", 1, 999.99},
             {"John Doe", "Tablet", 2, 599.98}};
 
-        int rowCount = 0;
+        size_t rowCount = 0;
         while (rs->next())
         {
             REQUIRE(rowCount < expectedResults.size());
@@ -349,7 +349,7 @@ TEST_CASE("MySQL LEFT JOIN operations", "[mysql_real_left_join]")
             {"Eva Wilson", std::nullopt, std::nullopt, std::nullopt},
             {"Frank Miller", std::nullopt, std::nullopt, std::nullopt}};
 
-        int rowCount = 0;
+        size_t rowCount = 0;
         while (rs->next())
         {
             REQUIRE(rowCount < expectedResults.size());
@@ -393,7 +393,7 @@ TEST_CASE("MySQL LEFT JOIN operations", "[mysql_real_left_join]")
             {6, "Eva Wilson"},
             {7, "Frank Miller"}};
 
-        int rowCount = 0;
+        size_t rowCount = 0;
         while (rs->next())
         {
             REQUIRE(rowCount < expectedResults.size());
@@ -430,7 +430,7 @@ TEST_CASE("MySQL LEFT JOIN operations", "[mysql_real_left_join]")
         // This should execute but return only NULL values for the right side
         auto rs = conn->executeQuery(query);
 
-        int rowCount = 0;
+        size_t rowCount = 0;
         while (rs->next())
         {
             REQUIRE_FALSE(rs->isNull("customer_id"));
