+++ libs/cpp_dbc/test/test_sqlite_real_inner_join.cpp
@@ -235,7 +235,7 @@ TEST_CASE("SQLite INNER JOIN operations", "[sqlite_real_inner_join]")
                 {4, "Alice Brown", 1007, 499.99},
                 {5, "Charlie Davis", 1008, 299.99}};
 
-            int rowCount = 0;
+            size_t rowCount = 0;
             while (rs->next())
             {
                 REQUIRE(rowCount < expectedResults.size());
@@ -274,7 +274,7 @@ TEST_CASE("SQLite INNER JOIN operations", "[sqlite_real_inner_join]")
                 {"John Doe", "Laptop", 1, 999.99},
                 {"John Doe", "Tablet", 2, 599.98}};
 
-            int rowCount = 0;
+            size_t rowCount = 0;
             while (rs->next())
             {
                 REQUIRE(rowCount < expectedResults.size());
@@ -310,7 +310,7 @@ TEST_CASE("SQLite INNER JOIN operations", "[sqlite_real_inner_join]")
                 {"Bob Johnson", "Laptop", 1, 999.99},
                 {"Alice Brown", "Smartphone", 1, 499.99}};
 
-            int rowCount = 0;
+            size_t rowCount = 0;
             while (rs->next())
             {
                 REQUIRE(rowCount < expectedResults.size());
